Implement frequent flyers calculation

+++ src/main/scala/com/quantexa/flightdata/FlightData.scala
@@ -2,8 +2,8 @@ package com.quantexa.flightdata
 
 import java.sql.Date
 
-case class FlightData(passengerId: Int,
-                      flightId: Int,
+case class FlightData(passengerId: Long,
+                      flightId: Long,
                       from: String,
                       to: String,
                       date: Date)
