+++ src/main/scala/com/quantexa/flightdata/Passenger.scala
@@ -1,5 +1,5 @@
 package com.quantexa.flightdata
 
-case class Passenger(passengerId: Int,
+case class Passenger(passengerId: Long,
                      firstName: String,
                      lastName: String)
